[Misc] Using the aliases of builtin types like np.int is deprecated

+++ examples/pytorch/hilander/utils/evaluate.py
@@ -40,7 +40,7 @@ def evaluate(gt_labels, pred_labels, metric="pairwise"):
         "evaluate with {}{}{}".format(TextColors.FATAL, metric, TextColors.ENDC)
     ):
         result = metric_func(gt_labels, pred_labels)
-    if isinstance(result, np.float):
+    if isinstance(result, float):
         print(
             "{}{}: {:.4f}{}".format(
                 TextColors.OKGREEN, metric, result, TextColors.ENDC
